requirements.txt
amqp==2.5.2
anyjson==0.3.3
Babel==2.7.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
DateTime==4.3
Django==2.2.8
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
phonenumbers==8.11.1
PyJWT==1.7.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
twilio==6.35.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
zope.interface==4.7.1
